refactor :: renaming
- 변수 이름을 Id을 Email로 변경
