+++ src/screens/auth/signup/steps/EmailStep.tsx
